lesson4(corrections)
