+++ database/factories/BuildingsFactory.php
@@ -1,22 +1,31 @@
 <?php
 
-namespace Database\Seeders;
+namespace Database\Factories;
 
-use App\Models\Roles;
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
-use Illuminate\Database\Seeder;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
 use App\Models\Buildings;
-use DB;
 
-class BuildingsTableSeeder extends Seeder
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Buildings>
+ */
+class BuildingsFactory extends Factory
 {
+    protected $model = Buildings::class;
     /**
-     * Run the database seeds.
+     * Define the model's default state.
      *
-     * @return void
+     * @return array<string, mixed>
      */
-    public function run()
+    public function definition()
     {
-        Buildings::factory()->create();
+        return [
+            'id_address_eas'=>1,
+            'base_quarter_code'=>'5505',
+            'ground_area_code'=>18,
+            'building_code'=>1,
+            'prim_code'=>10,
+            'id_user'=>1
+        ];
     }
 }
